+++ lang/rhseo-pot.js
@@ -1,0 +1,8 @@
+const wpPot = require('wp-pot');
+
+wpPot({
+  destFile: './lang/rhseo.pot',
+  domain: 'rhseo',
+  package: 'RH SEO',
+  src: '../**/*.php'
+});